modules/location_consumer/requirements.txt
psycopg2-binary==2.9.6
kafka-python==2.0.2